requirements.txt
crewai==0.175.0
langchain_openai==0.3.32
pandas==2.3.2
pydantic==2.11.7
pyodbc==5.2.0
python-dotenv==1.1.1
SQLAlchemy==2.0.43
